feat(npm): refactor agent convnetions towards export and unit tests
