+++ package.json
@@ -10,14 +10,14 @@
     "test": "npm run build && cross-env NODE_ENV=test nyc ./node_modules/.bin/mocha --timeout 10000 --require babel-core/register --require babel-polyfill --recursive ./dist/test",
     "semantic-release": "semantic-release pre && npm publish && semantic-release post",
     "sls-deploy": "npm run build && node ./node_modules/serverless/bin/serverless deploy -v",
-    "sls-undeploy": "serverless remove",
+    "sls-undeploy": "node ./node_modules/serverless/bin/serverless remove",
     "sls-logs-hook": "node ./node_modules/serverless/bin/serverless logs -f gtmGithubHook -t",
     "sls-logs-results": "node ./node_modules/serverless/bin/serverless logs -f gtmGithubResults -t",
-    "agent": "node dist/src/ghEventAgent/ghEventAgent.js",
-    "agent-build": "npm run build && node ./dist/src/ghEventAgent/sass.js",
+    "agent": "node dist/src/agent/startAgent.js",
+    "agent-build": "npm run build && node ./dist/src/agent/sass.js",
     "docker-build": "docker build -t github-task-manager --no-cache .",
     "docker-start": "docker-compose up -d",
-    "docker-hub-run": "docker pull wyvern8/github-task-manager:latest && docker run --env-file ./.env github-task-manager",
+    "docker-hub-run": "docker pull wyvern8/github-task-manager:latest && docker run --env-file ./.env wyvern8/github-task-manager:latest",
     "docker-local-logs": "docker-compose logs --tail=0 --follow",
     "docker-local-bounce": "npm run agent-build && docker-compose restart && npm run docker-local-logs",
     "k8s-create": "chmod 755 ./k8s.sh && ./k8s.sh create",
