+++ src/agent/startAgent.js
@@ -1,0 +1,6 @@
+import { Agent } from './Agent';
+
+(() => {
+    let agent = new Agent();
+    agent.start();
+})();
