+++ src/agent/templates/event.html
