--- src/lib/sample.js
@@ -1,7 +1,0 @@
-export class Sample {
-
-    test() {
-        return 'test';
-    }
-
-}
